requirements.txt
streamlit
numpy
plotly
paho-mqtt